# internal/adaptors/rest/dto/ingredient_create_response.go
package dto

type IngredientCreateResponse struct {
	IngredientId          string `json:"ingredientId"`
	IngredientName        string `json:"ingredientName"`
	IngredientDescription string `json:"ingredientDescription"`
	Unit                  int64  `json:"unit"`
}